+++ src/widget/toggle_button_widget.rs
@@ -35,6 +35,7 @@ pub struct ToggleButtonWidget {
     text_widget: TextWidget,
     selected: bool,
     active: bool,
+    widget_id: i32,
 }
 
 impl ToggleButtonWidget {
@@ -54,6 +55,7 @@ impl ToggleButtonWidget {
             text_widget,
             selected: false,
             active: false,
+            widget_id: 0,
         }
     }
 
@@ -167,6 +169,14 @@ impl Widget for ToggleButtonWidget {
         }
 
         None
+    }
+
+    fn set_widget_id(&mut self, widget_id: i32) {
+        self.widget_id = widget_id;
+    }
+
+    fn get_widget_id(&mut self) -> i32 {
+        self.widget_id
     }
 
     fn draw(&mut self, c: Context, g: &mut GlGraphics, clip: &DrawState) {